delete estado inc

+++ protected/controllers/EstadoController.php
@@ -120,14 +120,23 @@ class EstadoController extends Controller
 	public function actionDelete()
 	{
 
-		//Preguntar si los estados a borrar estan vacios (no tienen dispositivos asociados)
+		$num = "SELECT COUNT(id_disp) FROM dispositivos WHERE id_estado IN (".$_POST['data'].")";
+
+		if($num==0){
+			$sql = "DELETE FROM estados WHERE id_estado IN (".$_POST['data'].")";
+			try {
+				Yii::app()->db->createCommand($sql)->query();
+				echo "1,El(los) registro(s) se ha borrado";			
+			} catch (Exception $e) {
+				echo "3,Error: existen activos asociados con ese estado";
+			}
+		}else{
+			echo "3,Error: existen activos asociados con ese estado";
+		}
 		
-		$sql = "DELETE FROM estados WHERE id_estado IN (".$_POST['data'].")";
-		if(Yii::app()->db->createCommand($sql)->query())
-			echo "1";
-		else
-			echo "2";
 	}
+
+	
 
 	/**
 	 * Lists all models.
